Move http-enum
