+++ src/http/http-method.enum.ts
@@ -1,0 +1,8 @@
+export enum HttpMethod {
+  GET = 'get',
+  POST = 'post',
+  PUT = 'put',
+  PATCH = 'patch',
+  DELETE = 'delete',
+  OPTIONS = 'options',
+}
