+++ src/index.ts
@@ -5,5 +5,5 @@ export * from './injector/injection-token.class'
 export * from './http/request.class'
 export * from './http/response.class'
 export * from './router/router.class'
-export * from './router/http-method.enum'
+export * from './http/http-method.enum'
 export * from './application/app-config.interface'
