--- src/router/http-method.enum.ts
@@ -1,8 +1,0 @@
-export enum HttpMethod {
-  GET,
-  POST,
-  PUT,
-  PATCH,
-  DELETE,
-  OPTIONS,
-}
