+++ src/router/route.interface.ts
@@ -1,5 +1,5 @@
 import { IAction } from '../controller/action.interface'
-import { HttpMethod } from './http-method.enum'
+import { HttpMethod } from '../http/http-method.enum'
 
 export interface IRoute {
   path: string,
